Added section to application's README that gives information on usage of the Custom404Mixin.  Split tests for mixin into a separate test file by classname.  Minor PEP8 formatting fix in mixin.
